Refactor input handling via ModeService
